+++ configs/inst_colorization/inst-colorizatioon_cocostuff-instance_256x256.py
