+++ mmedit/models/editors/inst_colorization/__init__.py
@@ -1,8 +1,8 @@
 # Copyright (c) OpenMMLab. All rights reserved.
+from .colorization_net import (FusionGenerator, InstanceGenerator,
+                               SIGGRAPHGenerator)
 from .inst_colorization import InstColorization
-from .inst_colorization_net import (FusionGenerator, InstanceGenerator,
-                                    SIGGRAPHGenerator)
-from .inst_colorization_generator import InstColorizationGenerator
+from .inst_colorization_net import InstColorizationGenerator
 
 __all__ = [
     'InstColorization', 'SIGGRAPHGenerator', 'InstanceGenerator',
